+++ Packages/InspectCore/Tests/CoreTests/TLSClientHelloSNIExtractorTests.swift
@@ -108,11 +108,3 @@ private func buildClientHello(serverName: String?) -> [UInt8] {
 
     return record
 }
-
-private func uint16(_ value: Int) -> [UInt8] {
-    [UInt8((value >> 8) & 0xFF), UInt8(value & 0xFF)]
-}
-
-private func uint24(_ value: Int) -> [UInt8] {
-    [UInt8((value >> 16) & 0xFF), UInt8((value >> 8) & 0xFF), UInt8(value & 0xFF)]
-}
