+++ Packages/InspectCore/Tests/CoreTests/TLSTestHelpers.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+func uint16(_ value: Int) -> [UInt8] {
+    [UInt8((value >> 8) & 0xFF), UInt8(value & 0xFF)]
+}
+
+func uint24(_ value: Int) -> [UInt8] {
+    [UInt8((value >> 16) & 0xFF), UInt8((value >> 8) & 0xFF), UInt8(value & 0xFF)]
+}
